Mise backend plugin for installing pulumi plugins
This adds the initial implementation of a mise backend plugin for
pulumi. The initial use case is for installing pulumi plugins via mise.
You can currently do this with mise using things like the `github`
backend (e.g. `github:pulumi/pulumi-aws`), but that method has some
limitations. It installs the plugin as a single binary which it puts in
your PATH so things like `pulumi plugin ls` will not pick this up.

This backend plugin solves this by using `pulumi plugin install` to
install the plugin to `PULUMI_HOME` and then symlinking it to the mise
install location. This way anything that relies on PULUMI_HOME will
still work _and_ mise will setup the PATH to ensure these versions are
used.