BackendInstall should copy the install instead of symlinking
It looks like symlinking the install does not work in GitHub actions.
When the mise action creates/restores the cache the symlink is lost and
the tools is not available on the filesystem. This PR changes the
symlink to a copy so that the actual file is copied.

There still leaves an issue where the PULUMI_HOME directory is not
updated after a cache restore (or when running `mise install`) on new
projects. I'll solve this in a follow up PR.

**Additional Info**
- Before mise installs a tool, it first gathers information from the
  filesystem. It looks to see if a version of the tool is installed, and
  if it is it completely bypasses the plugin.
- `BackendInstall` is _only_ called if a new version is needed.
- `BackendListVersions` is _only_ called when:
  - `mise install` is called and _no_ version of the tool is installed.
  - `mise upgrade` is called
- `BackendExecEnv` is called in a lot of places. Anytime `activate`,
  `install`, `env`, etc are called.