Respect MISE_GITHUB_TOKEN (#8)
If the mise action gets a `github_token` input it's consumed downstream
as `MISE_GITHUB_TOKEN`. We weren't doing this, so these steps weren't
actually getting auth.

We also need to fix our ESC configuration while we're in here, because
we hadn't setup OIDC so it wasn't actually giving us pulumi-bot's token.


https://github.com/jdx/mise-action/blob/e3592f6c799b8b316d74be2762081c2688aa3b4a/src/index.ts#L177-L185