Add a mise env hook for pulumi and go versions
This extracts the environment plugin from
https://github.com/pulumi/ci-mgmt/pull/1852 into our existing vfox
plugin.

Originally I had hoped to distribute the plugin in-repo, to reduce the
number of things we have to version, but that will require some upstream
changes.

It can be used like so:

```
[plugins]
"vfox-pulumi" = "https://github.com/pulumi/vfox-pulumi"

[env._."vfox-pulumi"]
```

By default the plugin looks for a `go.mod` file next to `mise.toml` or
at the root of the git repo.

Custom module paths, which a lot of our projects (unfortunately) use,
can be specified like so:

```
[env._."vfox-pulumi"]
module_path = "provider"
```

This will export `PULUMI_VERSION_MISE` (if pkg is a dependency) and
`GO_VERSION_MISE`.

If no `go.mod` is found no environment variables are exported.